+++ internal/entities/feed/item.go
@@ -1,7 +1,7 @@
 package feed
 
 import (
-	"encoding/json"
+	"time"
 
 	"github.com/google/uuid"
 
@@ -11,11 +11,40 @@ import (
 )
 
 const (
-	DaoCreated          Event = "dao.created"
-	ProposalCreated     Event = "proposal.created"
-	ProposalVoteStarted Event = "proposal.voting.started"
-	ProposalVoteEnded   Event = "proposal.voting.ended"
+	DaoCreated                  Event = "dao.created"
+	ProposalCreated             Event = "proposal.created"
+	ProposalUpdated             Event = "proposal.updated"
+	ProposalUpdatedState        Event = "proposal.updated.state"
+	ProposalVotingStartsSoon    Event = "proposal.voting.starts_soon"
+	ProposalVotingStarted       Event = "proposal.voting.started"
+	ProposalVotingReachedQuorum Event = "proposal.voting.quorum_reached"
+	ProposalVotingFinishesSoon  Event = "proposal.voting.coming"
+	ProposalVotingEndsSoon      Event = "proposal.voting.ends_soon"
+	ProposalVotingEnded         Event = "proposal.voting.ended"
+
+	ActionNone                        ActionSource = ""
+	ActionDaoCreated                  ActionSource = "dao.created"
+	ActionDaoUpdated                  ActionSource = "dao.updated"
+	ActionProposalCreated             ActionSource = "proposal.created"
+	ActionProposalUpdated             ActionSource = "proposal.updated"
+	ActionProposalVotingStartsSoon    ActionSource = "proposal.voting.starts_soon"
+	ActionProposalVotingStarted       ActionSource = "proposal.voting.started"
+	ActionProposalVotingQuorumReached ActionSource = "proposal.voting.quorum_reached"
+	ActionProposalVotingEnded         ActionSource = "proposal.voting.ended"
 )
+
+var AallowedEvents = []Event{
+	DaoCreated,
+	ProposalCreated,
+	ProposalUpdated,
+	ProposalUpdatedState,
+	ProposalVotingStartsSoon,
+	ProposalVotingStarted,
+	ProposalVotingReachedQuorum,
+	ProposalVotingFinishesSoon,
+	ProposalVotingEndsSoon,
+	ProposalVotingEnded,
+}
 
 type Event string
 
@@ -25,7 +54,6 @@ type Item struct {
 	UpdatedAt    common.Time        `json:"updated_at"`
 	ReadAt       *common.Time       `json:"read_at"`
 	ArchivedAt   *common.Time       `json:"archived_at"`
-	Event        Event              `json:"event"`
 	DaoID        uuid.UUID          `json:"dao_id"`
 	ProposalID   string             `json:"proposal_id"`
 	DiscussionID string             `json:"discussion_id"`
@@ -33,5 +61,30 @@ type Item struct {
 	Action       string             `json:"action"`
 	DAO          *dao.DAO           `json:"dao,omitempty"`
 	Proposal     *proposal.Proposal `json:"proposal,omitempty"`
-	Timeline     json.RawMessage    `json:"timeline"`
+	Timeline     []Timeline         `json:"timeline"`
+}
+
+type Timeline struct {
+	CreatedAt common.Time `json:"created_at"`
+	Event     Event       `json:"event"`
+}
+
+// ActionSourceMap TODO: Move it to the SDK
+var ActionSourceMap = map[ActionSource]Event{
+	ActionDaoCreated:                  DaoCreated,
+	ActionProposalCreated:             ProposalCreated,
+	ActionProposalUpdated:             ProposalUpdated,
+	ActionProposalVotingStartsSoon:    ProposalVotingStartsSoon,
+	ActionProposalVotingStarted:       ProposalVotingStarted,
+	ActionProposalVotingQuorumReached: ProposalVotingReachedQuorum,
+	ActionProposalVotingEnded:         ProposalVotingEnded,
+}
+
+// ActionSource TODO: Move it to the SDK
+type ActionSource string
+
+// TimelineSource TODO: Move it to the SDK
+type TimelineSource struct {
+	CreatedAt time.Time    `json:"created_at"`
+	Action    ActionSource `json:"action"`
 }